+++ scripts/src/admin/CenterList/CenterList.d.ts
@@ -6,5 +6,5 @@ interface ICenterListState {
 }
 
 interface ICenterListProps {
-
+    onError: (any) => void
 }